refactor(admin-31): Rename all files (and changes paths)
