--- src/ui/components/forms/datefield.component.tsx
@@ -1,49 +1,0 @@
-import { FC, useEffect, ChangeEventHandler } from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { IconDefinition } from "@fortawesome/fontawesome-svg-core";
-import { faCalendar } from "@fortawesome/free-solid-svg-icons";
-
-import { dateToInputFormat } from "@/shared/utils/common/date.utils";
-
-import "@/ui/components/forms/textfield.component.css";
-
-interface DateFieldCompProps {
-    id           : string;
-    name         : string;
-    text         : string;
-    value?       : Date;
-    icon?        : IconDefinition;
-    onValueChange: ChangeEventHandler<HTMLInputElement>;
-}
-
-const DateFieldComp: FC<DateFieldCompProps> = ({ id, name, text, value = new Date(), icon = faCalendar, onValueChange }) => {
-    useEffect(() => {
-        console.log("Loaded: InfobarComp");
-    }, []);
-
-    useEffect(() => {
-        console.log("Rendered: InfobarComp");
-    });
-
-    return (
-        <div className="form-field form-field-date">
-            <FontAwesomeIcon icon={icon} />
-
-            <label htmlFor={id}>
-                {text}
-            </label>
-
-            <input
-                id={id}
-                name={name}
-                value={
-                    dateToInputFormat(value)
-                }
-                type="date"
-                onChange={onValueChange}
-            />
-        </div>
-    );
-};
-
-export default DateFieldComp;
