+++ src/ui/pages/admin/event/events_admin.page.tsx
@@ -1,23 +1,15 @@
 import { FC, ReactNode, useEffect, useState } from "react";
-import EventEditForm from "@/ui/components/admin/eventedit.component";
-import ClubSectionEdit from "@/ui/components/admin/clubsectionedit.component";
-import HorizontalSeparatorComp from "@/ui/components/forms/horizontalseparator.component";
-import FormBtnComp from "@/ui/components/forms/formbtn.component";
-import BoxDownFieldComp from "@/ui/components/forms/boxdownfield.component";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faHashtag } from "@fortawesome/free-solid-svg-icons";
+import EventForm from "@/ui/components/admin/event_edit.component";
+import HorizontalSeparatorComp from "@/ui/components/forms/horizontal_separator.component";
 
-import "@/ui/pages/admin/admin.page.css";
-import "@/ui/components/admin/eventedit.component.css";
-
-import { faClone, faTrash, faPen } from "@fortawesome/free-solid-svg-icons";
 import { getListEvents } from "@/api/event.api";
 import { Event } from "@/shared/models/event.model";
 
+import "@/ui/pages/admin/admin.page.css";
 
 const EventsAdminPage: FC = (): ReactNode => {
     const [events, setEvents] = useState<Event[] | undefined>(undefined);
-    
+
     useEffect(() => {
         getListEvents()
             .then(setEvents)
@@ -25,11 +17,13 @@ const EventsAdminPage: FC = (): ReactNode => {
     }, []);
 
     return (
-        <div className="clubs-modification-page">
+        <div className="event-modification-page">
             
-            <h2>Clubs Modication Page (Admin Page)</h2>
+            <h2>Events Modication Page (Admin Page)</h2>
 
             <text>
+                <p>Select an event either from the list on the left or from the “Selected Event” section. </p>
+
                 <p>
                     Nemo quaeso miretur, si post exsudatos labores itinerum longos congestosque adfatim commeatus
                     fiducia vestri ductante barbaricos pagos adventans velut mutato repente consilio ad placidiora deverti.
@@ -46,36 +40,13 @@ const EventsAdminPage: FC = (): ReactNode => {
 
             <HorizontalSeparatorComp />
 
-            <div className="event-form" style={{ width: "100%" }}>
-                <BoxDownFieldComp text="Club Select" values={["BDE", "CIA", "PolyChrome"]} icon={faPen} />
-                
-                <div className="col3">
-                    <FormBtnComp text="New Event" />
-                    <FormBtnComp text="Duplicate" icon={faClone} />
-                    <FormBtnComp text="Delete" icon={faTrash} />
-                </div>
-            </div>
-            
-            <HorizontalSeparatorComp />
-
-            <EventEditForm
+            <EventForm
                 events={events ?? []}
                 onSave={async (updated_event) => {
                     console.log("Saving event:", updated_event);
                     await Promise.resolve();
                 }}
             />
-
-            <h3>
-                <FontAwesomeIcon icon={faHashtag} /> Event Preview
-            </h3>
-
-            <HorizontalSeparatorComp />
-            <ClubSectionEdit />
-
-            <h3>
-                <FontAwesomeIcon icon={faHashtag} /> Section Preview
-            </h3>
         </div>
     );
 };
